requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
SQLAlchemy>=2.0
psycopg2-binary
Flask-Migrate>=4.0
python-dotenv
Flask-Bcrypt
PyJWT

flask-restx>=1.3.0
pytest
debugpy

Werkzeug==2.2.2
Flask-SQLAlchemy==3.1.0
SQLAlchemy==2.0.16
psycopg2-binary==2.9.5
Flask-Migrate==4.0.7
python-dotenv==0.21.0
sqlalchemy-utils==0.37.8
jsonschema>=4.18.0